==> code/analysis/04_fgsea.Rmd <==
---
title: "04_GSEA analysis"
author: "Nick Diercksen"
date: "Last compiled on `r format(Sys.time())`"
output:
  html_notebook:
    toc: yes
    toc_float: yes
  html_document:
    toc: yes
    df_print: paged
editor_options:
  chunk_output_type: inline
params:
  FCutoff:
    label: cutoff for the log2FoldChange
    value: 1
  pCutoff:
    label: "cutoff for the adjusted p-Value"
    value: 0.01
  fgsea.maxSize:
    label: "parameter for `fgsea`: 'Maximal size of a gene set to test. All pathways above the threshold are excluded.'"
    value: 200
  reevaluate:
    label: Should all calculations and database lookups be done again or just e.g.
      graphics regenerated
    value: no
  condition_pal: !r c("WT"= "#00C3C6","KO"= "#FF6C67")
  tissue_pal: !r c(gastroc = "orange", soleus="purple")
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(fgsea)

# plotting functions
library(ggplot2)
library(eulerr)

library(patchwork) # (combining ggplots with '+' => needed for the dotplots, ranks)
library(ggpubr)   # arrange plots (top enrichment)
```


```{r plottingSettings, include=FALSE}
# ggplot2::theme_set(theme_minimal())
ggplot2::theme_set(theme_bw())

plot.dir <- file.path("plots", "04 GSEA")
```


In this document the Gene Set Enrichment Analysis (GSEA) will be performed using
`fgsea`. Besides the package, the results from the DEA are needed to provide the
ranked list of genes to be investigated for enrichment, and the pathways stemming
from the previous step.


```{r load_rdata}
load("./data/Robjects/02_DDS.RData")   # for `stat` values from DESeq results
load("./data/Robjects/03_annot.RData") # for entrezgene_id
```

# ranks

## obtaining ranks

the `ranks` will be a `named num`, with `entrezgene_id` as name and `stat` (Wald
Test) as metric.

Currently all genes for which Entrez IDs could be found and were investigated in
the DEA are used in this analysis. Meaning independent of whether they were flagged
as significantly differentially expressed.
!This step here could be adjusted in the future to only contain genes, which were
flagged as significantly DE.

```{r getting_ranks}
getRanks <- function(res, annot) {
  # only taking genes which have entrezgene_ids assigned to them
  genes_with_entrez <- select(annot, GeneID, entrezgene_id) %>% 
    filter(!is.na(entrezgene_id))
  
  ranks <- as.data.frame(res) %>%
    tibble::rownames_to_column("GeneID") %>%
    merge(genes_with_entrez, by = "GeneID") %>%
    arrange(desc(stat)) %>% 
    select(entrezgene_id, stat) %>% 
    tibble::deframe() # creating a named num from two columns
  return(ranks)
}

ranks.gastroc <- getRanks(res.gastroc, annot)
ranks.soleus <- getRanks(res.soleus, annot)
```


## distribution of ranks

As the t-statistics represent the ranks, their distribution is plotted for both
tissues.

```{r ranks_barplot, echo=FALSE}
plot_ranks <- function(ranks, title) {
  reshape2::melt(ranks, value.name = "rank") %>%
    ggplot(aes(x=seq_along(rank),y=rank)) +
    geom_bar(stat="identity") +
    ylim(c(-20,20)) +
    labs(x = "gene rank in ordered dataset", y = "Wald Test value") +
    ggtitle(label = title)
}

p.gastroc <- plot_ranks(ranks.gastroc, "gastrocnemius")
p.soleus <- plot_ranks(ranks.soleus, "soleus")

patchwork::wrap_plots(p.gastroc, p.soleus) + 
  patchwork::plot_annotation(title = "rank distributions")

```


The distribution clearly resembles the volcano or MA plots of the DEA in step 2,
as the genes of the gastrocnemius show significantly higher magnitudes.

The x-axis of the gastrocnemius is also slightly bigger. With `r length(ranks.gastroc)` genes the gastrocnemius gene count is more than  300 over the one from the
soleus (`r length(ranks.soleus)`). This is due to the pre-filtering in step 1.


# pathways

Pathways are provided by <http://www.gsea-msigdb.org/gsea/msigdb/mouse/collections.jsp>

For now the Canonical pathways are used. These gene sets represent biological a biological process. They are composed from the following databases taking a subset of CP:

| database     | gene sets |
|--------------|-----------|
| BioCarta     | 252       |
| Reactome     | 1249      |
| WikiPathways | 186       |

```{r reading_pathways, include=FALSE}
CGP <- qusage::read.gmt("./data/pathways/m2.cp.v2022.1.Mm.entrez.gmt") # canonical pathways (1687 gene sets)
```

# applying fgsea

`fgsea` is called as depicted below.


    fgseaRes <- fgsea(
      pathways = CGP,
      stats    = ranks,
      minSize  = 15,
      maxSize  = `r params$fgsea.maxSize`
    )


The actual used code was constructed so that it was easier to adjust the maximum
size of a pathway (number of genes). It was a parameter to be tuned to yield
yield proper results:

```{r fgsea_res_load}
# ' returns the fgsea result
# ' `tissue` needs to be one of {"soleus", "soleus"}
# ' It saves the results locally to avoid too many queries
get_fgseaRes <- function(tissue, ranks, maxSize=params$fgsea.maxSize, pathways=CGP, saveResults=T) {
  robjects.path <- file.path(".", "data", "Robjects")
  fgseaRes.path <- file.path(
    robjects.path,
    "04_fgsea_results",
    tissue,
    paste0("maxSize_", params$fgsea.maxSize, ".rds")
  )
  fgseaRes.dir <- dirname(fgseaRes.path)
  
  if (file.exists(fgseaRes.path) & !params$reevaluate) {
    message("fgsea result found in directory! Previous version will be loaded ...")
    return(readRDS(file = fgseaRes.path))
  } else {
    # perform fgsea and save the result
    fgseaRes <- fgsea(
      pathways = pathways,
      stats    = ranks,
      minSize  = 15,
      maxSize  = maxSize
    )
    
    if (saveResults) {
      if (!dir.exists(fgseaRes.dir))
        dir.create(fgseaRes.dir, recursive = T)
      saveRDS(fgseaRes, file = fgseaRes.path)
    }
    return(fgseaRes)
  }
}

fgseaRes.gastroc <- get_fgseaRes("gastroc", ranks.gastroc, pathways=CGP)
fgseaRes.soleus <- get_fgseaRes("soleus", ranks.soleus, pathways=CGP)
```

# most differential regulated pathways, both tissues

The most significantly enriched pathways are chosen by having a adjusted p-value
higher than `r params$pCutoff`.

Taking these Normalized Enrichment Scores (NES) from both tissues, they can be
categorized whether both tissues are up- or down-regulated or just one…

```{r filter_most_diff_pathways}
pCutoff <- params$pCutoff

fgseaRes.combined <- merge(
  data.frame(fgseaRes.gastroc[, c("pathway", "NES", "padj")]),
  data.frame(fgseaRes.soleus[, c("pathway", "NES", "padj")]),
  by = "pathway",
  suffixes = c(".ga", ".sol")
) %>%
  filter(padj.ga < pCutoff | padj.sol < pCutoff) %>%
  mutate(
    diff.exp = case_when(
      NES.ga  < 0 & NES.sol < 0 & padj.ga < pCutoff & padj.sol < pCutoff ~ "both down",
      NES.ga  > 0 & NES.sol > 0 & padj.ga < pCutoff & padj.sol < pCutoff ~ "both up",
      NES.ga  < 0 & NES.sol > 0 & padj.ga < pCutoff & padj.sol < pCutoff ~ "gastrocnemius down, soleus up",
      NES.ga  > 0 & NES.sol < 0 & padj.ga < pCutoff & padj.sol < pCutoff ~ "gastrocnemius up, soleus down",
                                  padj.ga < pCutoff & padj.sol > pCutoff ~ "only gastrocnemius",
      # NES.ga  > 0 &               padj.ga < pCutoff & padj.sol > pCutoff ~ "ga up",
                                  padj.ga > pCutoff & padj.sol < pCutoff ~ "only soleus",
      # NES.sol > 0 &               padj.ga > pCutoff & padj.sol < pCutoff ~ "sol up",
      TRUE ~ "different"
    )
  )
```

## gsea scatter

using `NES` from the fgsea result filtering on the set ``` pCutoff=``r params$pCutoff ``` yields the following plot:

```{r topDiff.both, warning=FALSE}
p <- ggplot(fgseaRes.combined, aes(x = NES.ga, y = NES.sol, text=pathway)) +
  geom_vline(xintercept = 0) + 
  geom_hline(yintercept = 0) + 
  geom_point(aes(color = diff.exp)) +
  scale_color_manual(values = c("royalblue", "red", unname(params$tissue_pal["gastroc"]), unname(params$tissue_pal["soleus"]))) +
  labs(x = "gastrocnemius", y = "soleus", color = "significantly\ndifferentially\nexpressed") +
  # ggrepel::geom_label_repel(max.overlaps = 20) + 
  ggtitle(label = "Normalized Enrichment Score")
```


```{r topDiff.both_save, warning=FALSE, include=FALSE}
# save plain version
ggsave(filename = file.path("plots", "04_gsea_scatter.svg"), p)
ggsave(filename = file.path("plots", "04_gsea_scatter.png"), p)

# display interactive version
plotly::ggplotly(p, tooltip = "all")
```


barplot:

```{r gsea_scatter_barplot, include=FALSE}
# belongs to the scatter plot for better visualization of the distribution
p <- ggplot(fgseaRes.combined, aes(x = diff.exp, fill = diff.exp)) +
  scale_fill_manual(values = c("royalblue", "red", unname(params$tissue_pal["gastroc"]), unname(params$tissue_pal["soleus"]))) +
  geom_bar(aes(fill = diff.exp))

ggsave(filename = file.path("plots", "04_gsea_scatter_barplot.svg"), p)
ggsave(filename = file.path("plots", "04_gsea_scatter_barplot.png"), p)
p
```

Euler-Diagram:

```{r eulerDiagramGSEA}
venn.colors <- c(params$tissue_pal, scales::hue_pal()(4))

sign_geneSets_gastroc <-
  data.frame(fgseaRes.gastroc[, c("pathway", "padj")]) %>%
  filter(padj < params$pCutoff) %>% 
  pull(pathway)


sign_geneSets_soleus <-
  data.frame(fgseaRes.soleus[, c("pathway", "padj")]) %>%
  filter(padj < params$pCutoff) %>% 
  pull(pathway)

# only the two tissues
gene_sets <- list(
  "gastrocnemius" = sign_geneSets_gastroc,
  "soleus" = sign_geneSets_soleus
)

# euler: two tissues
p <- plot(
  euler(gene_sets),
  quantities = list(cex = 2),
  legend = list(side = "bottom", cex = 2),
  fills = venn.colors,
  # main = "significant gene sets (GSEA)"
)
ggsave(filename = file.path(plot.dir, "04_euler.svg"), p)
ggsave(filename = file.path(plot.dir, "04_euler.png"), p)
p
```



## dotplot

For better visualization of the GSEA results, the significant genes are presented
with dotpltos.

Here a custom dotplot function is created. To compare the findings in both tissues,
two dotplots, one for each tissue, can be created side by side.

```{r echo=FALSE, include=FALSE}
# ' Creates a dotplot for each tissue of the respective group.
# ' ----------
# ' The group needs to be value within the `diff.exp` column of the 
# ' `fgseaRes.combined`.
# ' 
combined_dotplot <-
  function(group,
           fgseaRes.combined = fgseaRes.combined,
           max_label_length = 25, topn=6e6, sort_by_gastroc=T) {
    
    # get pathways of the respective group
    pathways <- filter(fgseaRes.combined, diff.exp == group)$pathway
    
    gastroc.df <-
      filter(fgseaRes.gastroc, pathway %in% pathways)
    soleus.df <-
      filter(fgseaRes.soleus, pathway %in% gastroc.df$pathway)
    
    if (sort_by_gastroc) {
      gastroc.df <- gastroc.df %>% 
        top_n(topn, wt = abs(NES)) %>%
        arrange(NES)
      soleus.df <- soleus.df %>% 
        .[match(gastroc.df$pathway, pathway)]
    } else {
      soleus.df <- soleus.df %>% 
        top_n(topn, wt = abs(NES)) %>%
        arrange(NES)
      gastroc.df <- gastroc.df %>% 
        .[match(soleus.df$pathway, pathway)]
    }
	
    padj_limits <-
      c(min(gastroc.df$padj, soleus.df$padj),
        max(gastroc.df$padj, soleus.df$padj))
    
    NES_limits <- 
      c(min(gastroc.df$NES, soleus.df$NES),
        max(gastroc.df$NES, soleus.df$NES))
    
    p.gastroc <- fgsea_dotplot(gastroc.df, max_label_length = max_label_length, padj_limits, NES_limits)

    p.soleus <- fgsea_dotplot(soleus.df, max_label_length = max_label_length, padj_limits, NES_limits)
    
    # adjust plots
    p.gastroc <- p.gastroc +
      theme(
        legend.position = "none",
        plot.margin = unit(c(0, 0, 0, 0), "npc"),
      ) +
      ylab("NES\n(gastroc)") + 
      xlab("")
    p.soleus <- p.soleus +
      theme(
        axis.text.y = element_blank(),
        plot.margin = unit(c(0, 0, 0, 0), "npc")
      ) +
      ylab("NES\n(soleus)") + 
      xlab("")
    
    return(p.gastroc + p.soleus)
  }


# ' Creates a dotplot of the given results dataframe
# ' ----------
fgsea_dotplot <-
  function(fgseaRes.df,
           max_label_length=25,
           padj_limits=NULL,
           NES_limits=NULL) {
  size_breaks <- seq(0,200,by=50)
  size_breaks[1] <- 10
  size_range <- c(2,8)
  
  label_size <- 8
  nsets <- nrow(fgseaRes.df)
  if (nsets > 15) {
    # adjust label size if the plot has too many sets:
    label_size <- label_size - (0.05 * nsets)
    max_label_length <- nsets + 10
    size_range <- size_range -1
  }
  
  fgseaRes_prep.df <- fgseaRes.df %>%
    # TODO: might need to be changed if other labels are used (like description)
    mutate(pathway = gsub("_", " ", pathway) %>% tolower()) %>%
    mutate(pathway = factor(pathway, levels = pathway))
  # the factor levels might be different than the order of the rows
  
  ggplot(fgseaRes_prep.df, aes(x = pathway, y = NES)) +
    geom_point(aes(size = size, color = padj)) +
    scale_color_gradient(low = "red",
                         high = "blue",
                         limits = padj_limits) +
    scale_size_continuous(range = size_range,
                          breaks = size_breaks,
                          limits = c(10, 200)) +
    scale_y_continuous(limits = NES_limits) +
    # scale_x_discrete(guide = guide_axis(check.overlap = TRUE))+
    scale_x_discrete(labels = scales::label_wrap(max_label_length)) +
    xlab("Gene Set") + ylab("NES") +
    labs(size = "Count", color = "p.adjust") +
    theme_bw() +
    coord_flip() +
    theme(axis.text.y = element_text(size = label_size))
  }
```

> Note: Different to most other plots, the dotplots seem to be created best by
running the chunks manually. Especially the plots saved to the file-system.
This results in automatic axis scaling with more entries, so that the labels do
not overlap (which is curiously not the case if remotely knitting)!


In the following the dotplots are created for the different groups identified
above …


### "both upregulated"
only 7 gene sets => fits well in the plot
```{r dotplotBothUp}
group = "both up"
p_combined <- combined_dotplot(group, fgseaRes.combined)
p_combined
```

```{r dotplotBothUp_save, include=FALSE}
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, ".svg")), p_combined)
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, ".png")), p_combined)
```


### "both down"
gets adjusted by the function well. The .svg file looks okay
```{r dotplotBothDown}
group = "both down"
p_combined <- combined_dotplot(group, fgseaRes.combined)
p_combined
```

```{r dotplotBothDown_save, include=FALSE}
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, ".svg")), p_combined)
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, ".png")), p_combined)
```


### "only gastrocnemius"

```{r dotplotGastroc}
group = "only gastrocnemius"
p_combined <- combined_dotplot(group, fgseaRes.combined, max_label_length = 200)
p_combined
```

```{r dotplotGastroc_save, include=FALSE}
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, ".svg")), p_combined)
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, ".png")), p_combined)
```

The results of gastrocnemius in a single plot:

```{r dotplotGastrocSingle}
group = "only gastrocnemius"
pathways <- filter(fgseaRes.combined, diff.exp == group)$pathway

gastroc.df <-
  filter(fgseaRes.gastroc, pathway %in% pathways) %>%
  arrange(NES)

p.gastroc <- fgsea_dotplot(gastroc.df, max_label_length = 100)
p.gastroc
```

```{r dotplotGastrocSingle_save, include=FALSE}
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, "_single.svg")), p.gastroc)
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, "_single.png")), p.gastroc)
```


### "only soleus"

```{r dotplotSoleus}
group = "only soleus"
p_combined <- combined_dotplot(group, fgseaRes.combined, max_label_length = 100, sort_by_gastroc = F)
p_combined
```

```{r dotplotSoleus_save, include=FALSE}
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, ".svg")), p_combined)
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, ".png")), p_combined)
```


The results of soleus in a single plot:

```{r dotplotSoleusSingle}
group = "only soleus"
pathways <- filter(fgseaRes.combined, diff.exp == group)$pathway

soleus.df <-
  filter(fgseaRes.soleus, pathway %in% pathways) %>%
  arrange(NES)

p.soleus <-
  fgsea_dotplot(soleus.df, max_label_length = 100)
p.soleus
```

```{r dotplotSoleusSingle_save, include=FALSE}
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, "_single.svg")), p.soleus)
ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, "_single.png")), p.soleus)
```




```{r top_20 eval=FALSE, include=FALSE}
# could
for (group in unique(fgseaRes.combined$diff.exp)) {
  p_combined <- combined_dotplot(group, fgseaRes.combined, topn = 20)
  
  ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, "top20", ".svg")), p_combined)
  ggsave(filename = file.path("plots", paste0("04_gsea_dotplot_", group, "top20", ".png")), p_combined)
  print(p_combined)
}
```


# Enrichment score plot

ordering pathways by padj values and using `ES` to

```{r fn_get_top_pathways}
# ' obtain top pathways ordered by padj and use `ES` for up or down regulation
get_top_pathways <- function(fgseaRes, up = TRUE, pCutoff=params$pCutoff, n=10) {
  .updown <- ifelse(up, `>`, `<`)
  
  top.pathways <- fgseaRes %>%
    filter(.updown(ES,0), padj < pCutoff) %>%
    arrange(padj) %>% 
    slice_head(n=n)
  
 return(top.pathways) 
}
```

plot for top up and down regulated pathways

```{r}
# ' plots top n enrichment plots for the given fgsea result
plot_top_enrichment <- function(fgseaRes, pathways, ranks, n = 9, up = TRUE) {
  # extracting the top n pathways
  top.pathways <- get_top_pathways(fgseaRes, up=up, pCutoff=params$pCutoff, n=n)
  
  plot.list <- list()
  # lims <- list("x" = c(0,17000), "y" = c(-0.8,0.0))
  
  for (i in 1:nrow(top.pathways)) {
    # filling plot.list with enrichmentPlots 
    # TODO: how can I use facet_wrap for this?
    pathway <- top.pathways[i]$pathway
    plt <- plotEnrichment(pathways[[pathway]], ranks) +
      # TODO: adjust yaxis to the same scale
      # TODO: keep axis.text.x only on the lower row
      # TODO: keep axis.text.y only on the right column
      theme(
        axis.title.x = element_blank(),
        axis.title.y = element_blank()
      ) # +
      # coord_cartesian(xlim = lims$x, ylim = lims$y)
    plot.list[[i]] <- plt
  }
    
  arrange_plts(plot.list)
}

# ' helper function to arrange the plot from the enrichment
arrange_plts <- function(plt.list) {
  nplts <- length(plt.list)
  plt <- plt.list[1]
  xlab <- plt$labels$x
  ylab <- plt$labels$y
  
  # set axis to the same scale
  lims <- list("x" = c(0, 17000), "y" = c(-0.8, 0.0))
  
  # remove axis
  
  # arrange the plots
  fig_labels <- LETTERS[1:nplts]
  
  patchwork::wrap_plots(plt.list, )
  
  figure <- ggpubr::ggarrange(plotlist = plt.list,
                              labels = fig_labels) %>%
    annotate_figure(left = text_grob(ylab, rot = 90),
                    bottom = text_grob(xlab))
 
  # TODO: remove all x-axis labels except lower row
  # get dimensions
  figure$layers
   
  return(figure)
}



# plot_labels <-
#     data.frame("label" = LETTERS[1:10], "pathway" = top.pathways)
# knitr::kable(caption = "plot labels", plot_labels)

```

## gastroc up

```{r}
plot_top_enrichment(fgseaRes.gastroc, CGP, ranks.gastroc, up=T)

# TODO: add plot labels to return argument of plot_top_enrichment (use list probably)
plot_labels <-
    data.frame("label" = LETTERS[1:9], "pathway" = get_top_pathways(fgseaRes.gastroc, up=T, n=9)$pathway)
knitr::kable(caption = "plot labels", plot_labels)
```

## gastroc down

```{r}
plot_top_enrichment(fgseaRes.gastroc, CGP, ranks.gastroc, up=F)
plot_labels <-
    data.frame("label" = LETTERS[1:9], "pathway" = get_top_pathways(fgseaRes.gastroc, up=F, n=9)$pathway)
knitr::kable(caption = "plot labels", plot_labels)
```

## soleus up

```{r}
plot_top_enrichment(fgseaRes.soleus, CGP, ranks.soleus, up=T)

plot_labels <-
    data.frame("label" = LETTERS[1:9], "pathway" = get_top_pathways(fgseaRes.soleus, up=T, n=9)$pathway)
knitr::kable(caption = "plot labels", plot_labels)
```

## soleus down

```{r}
plot_top_enrichment(fgseaRes.soleus, CGP, ranks.soleus, up=F)
plot_labels <-
    data.frame("label" = LETTERS[1:9], "pathway" = get_top_pathways(fgseaRes.soleus, up=F, n=9)$pathway)
knitr::kable(caption = "plot labels", plot_labels)
```


## opposite regulated

```{r enrichmentPlotOpposite}
plotEnrichment(CGP[["REACTOME_EUKARYOTIC_TRANSLATION_INITIATION"]], ranks.gastroc)
plotEnrichment(CGP[["REACTOME_EUKARYOTIC_TRANSLATION_INITIATION"]], ranks.soleus)
```



# GSEA table plot

GESEA table plots of the top 10 up- and down-regulated pathways.

## gastroc


```{r}
# creating up and down regulated pathway vectors separately to maintain order

topUp <- get_top_pathways(fgseaRes.gastroc, up=T, pCutoff = params$pCutoff, n=10)
topDown <- get_top_pathways(fgseaRes.gastroc, up=F, pCutoff = params$pCutoff, n=10)
topPathways <- bind_rows(topUp, topDown) %>%
  arrange(-NES) %>%
  pull(pathway)

plotGseaTable(
  pathways = CGP[topPathways],
  stats = ranks.gastroc,
  fgseaRes = fgseaRes.gastroc,
  gseaParam = 0.5,
  render = FALSE
) %>%
  ggpubr::as_ggplot() # needed since, for whatever reason only `NULL` gets returned if `plotGseaTable` is rendered inline
```

## soleus

top significant pathways:

```{r}
# creating up and down regulated pathway vectors separately to maintain order

topUp <- get_top_pathways(fgseaRes.soleus, up=T, pCutoff = params$pCutoff, n=10)
topDown <- get_top_pathways(fgseaRes.soleus, up=F, pCutoff = params$pCutoff, n=10)
topPathways <- bind_rows(topUp, topDown) %>%
  arrange(-NES) %>%
  pull(pathway)

plotGseaTable(
  pathways = CGP[topPathways],
  stats = ranks.soleus,
  fgseaRes = fgseaRes.soleus,
  gseaParam = 0.5,
  render = FALSE
) %>%
  ggpubr::as_ggplot() # needed since, for whatever reason only `NULL` gets returned if `plotGseaTable` is rendered inline
```

## list of all significant pathways

pathways will be sorted here by the magnitude of the NES…

### both down

```{r}
fgseaRes.combined %>% 
  filter(diff.exp == "both down") %>% 
  mutate(NES = abs(NES.ga) + abs(NES.sol)) %>% 
  select(pathway, NES) %>% 
  arrange(desc(NES)) %>%
  rmarkdown::paged_table(options = list(cols.min.print = 2))
```

### both up

```{r}
fgseaRes.combined %>% 
  filter(diff.exp == "both up") %>% 
  mutate(NES = abs(NES.ga) + abs(NES.sol)) %>% 
  select(pathway, NES) %>% 
  arrange(desc(NES)) %>%
  rmarkdown::paged_table(options = list(cols.min.print = 2))
```

### only gastroc

```{r}
fgseaRes.combined %>% 
  filter(diff.exp == "only gastrocnemius") %>% 
  mutate(NES = abs(NES.ga) + abs(NES.sol)) %>% 
  select(pathway, NES) %>% 
  arrange(desc(NES)) %>%
  rmarkdown::paged_table(options = list(cols.min.print = 2))
```

### only soleus

```{r}
fgseaRes.combined %>% 
  filter(diff.exp == "only soleus") %>% 
  mutate(NES = abs(NES.ga) + abs(NES.sol)) %>% 
  select(pathway, NES) %>% 
  arrange(desc(NES)) %>%
  rmarkdown::paged_table(options = list(cols.min.print = 2))
```

```{r include = FALSE}
saveRDS(fgseaRes.combined, file = "./data/Robjects/04_fgseaRes.combined.rds")
```
